edited how tag is set and read
Discovered that the "ID" on most cards and tags are in decimal format and only take 4 bytes, whereas the data obtained from the ID3-LA reader is not. In the SetTag() function, it now converts the original string provided by the user to a hex integer. Using stringstream, the hex integer is converted to a hex string with appropriate leading zeros to provide the correct number of characters. This string is what is used to compare the data read from the reader.